+++ hercules-ci-api-agent/test/Hercules/API/Agent/LifeCycle/AgentInfoSpec.hs
@@ -6,9 +6,9 @@ import Hercules.API.Agent.LifeCycle.AgentInfo
 import Test.Hspec
 import Prelude
 
-objectV1, objectV2 :: AgentInfo
+objectV1, objectV2, objectV3 :: AgentInfo
 
-jsonV1, jsonV2 :: BL.ByteString
+jsonV1, jsonV2, jsonV3 :: BL.ByteString
 jsonV1 = "{\"platforms\":[\"riscv-sel4\"],\"hostname\":\"a\",\"nixVersion\":\"nv\",\"agentVersion\":\"v\",\"cachixPushCaches\":[\"c1\",\"c2\"],\"systemFeatures\":[\"f1\",\"f2\"],\"substituters\":[\"s1\",\"s2\"]}"
 
 objectV1 = AgentInfo
@@ -18,6 +18,7 @@ objectV1 = AgentInfo
     platforms = ["riscv-sel4"],
     systemFeatures = ["f1", "f2"],
     cachixPushCaches = ["c1", "c2"],
+    pushCaches = [],
     substituters = ["s1", "s2"],
     concurrentTasks = 2 -- the hardcoded default
   }
@@ -31,6 +32,21 @@ objectV2 = AgentInfo
     platforms = ["riscv-sel4"],
     systemFeatures = ["f1", "f2"],
     cachixPushCaches = ["c1", "c2"],
+    pushCaches = [],
+    substituters = ["s1", "s2"],
+    concurrentTasks = 19 -- something else
+  }
+
+jsonV3 = "{\"platforms\":[\"riscv-sel4\"],\"hostname\":\"a\",\"nixVersion\":\"nv\",\"agentVersion\":\"v\",\"cachixPushCaches\":[\"c1\",\"c2\"],\"pushCaches\":[\"pc1\",\"pc2\"],\"systemFeatures\":[\"f1\",\"f2\"],\"substituters\":[\"s1\",\"s2\"], \"concurrentTasks\": 19}"
+
+objectV3 = AgentInfo
+  { hostname = "a",
+    agentVersion = "v",
+    nixVersion = "nv",
+    platforms = ["riscv-sel4"],
+    systemFeatures = ["f1", "f2"],
+    cachixPushCaches = ["c1", "c2"],
+    pushCaches = ["pc1", "pc2"],
     substituters = ["s1", "s2"],
     concurrentTasks = 19 -- something else
   }
@@ -40,7 +56,12 @@ spec = describe "AgentInfo" $ do
   describe "FromJSON" $ do
     it "parses v1 correctly" $ eitherDecode jsonV1 `shouldBe` Right objectV1
     it "parses v2 correctly" $ eitherDecode jsonV2 `shouldBe` Right objectV2
+    it "parses v3 correctly" $ eitherDecode jsonV3 `shouldBe` Right objectV3
   describe "ToJSON"
     $ it "encodes v2 correctly"
     $ eitherDecode (encode objectV2)
       `shouldBe` Right objectV2
+  describe "ToJSON"
+    $ it "encodes v3 correctly"
+    $ eitherDecode (encode objectV3)
+      `shouldBe` Right objectV3